Renamed filenames to be snakecase. Updated the contributing readme, and changed opensource9ja to javascriptdata
