+++ package.json
@@ -18,14 +18,14 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/opensource9ja/scikit.js.git"
+    "url": "git+https://github.com/javascriptdata/scikit.js.git"
   },
   "author": "",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/opensource9ja/scikit.js/issues"
+    "url": "https://github.com/javascriptdata/scikit.js/issues"
   },
-  "homepage": "https://github.com/opensource9ja/scikit.js#readme",
+  "homepage": "https://github.com/javascriptdata/scikit.js#readme",
   "devDependencies": {
     "@rollup/plugin-commonjs": "^21.0.1",
     "@rollup/plugin-node-resolve": "^13.0.6",
