+++ scikitjs-browser/package.json
@@ -68,7 +68,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/opensource9ja/scikit.js.git"
+    "url": "git+https://github.com/javascriptdata/scikit.js.git"
   },
   "keywords": [
     "scikit-learn",
@@ -79,9 +79,9 @@
   "author": "Rising Odegua <risingodegua@gmail.com>",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/opensource9ja/scikit.js/issues"
+    "url": "https://github.com/javascriptdata/scikit.js/issues"
   },
-  "homepage": "https://github.com/opensource9ja/scikit.js#readme",
+  "homepage": "https://github.com/javascriptdata/scikit.js#readme",
   "nyc": {
     "reporter": [
       "lcov",
