+++ shared/lib/dummy/dummyRegressor.ts
