+++ shared/lib/estimators/linearRegression.test.ts
@@ -1,6 +1,6 @@
 import * as tf from '@tensorflow/tfjs-core'
 import { assert } from 'chai'
-import { LinearRegression } from './linear.regression'
+import { LinearRegression } from './linearRegression'
 import { tensorEqual } from '../utils'
 import { Tensor1D, Tensor2D } from '@tensorflow/tfjs-core'
 import { describe, it } from 'mocha'
