+++ shared/lib/estimators/logisticRegression.test.ts
@@ -1,6 +1,6 @@
 import '@tensorflow/tfjs-backend-webgl'
 import { assert } from 'chai'
-import { LogisticRegression } from './logistic.regression'
+import { LogisticRegression } from './logisticRegression'
 import { arrayTo2DColumn } from '../utils'
 import { describe, it } from 'mocha'
