+++ shared/lib/metrics/metrics.ts
@@ -45,7 +45,7 @@ function assertInputIsWellFormed(labels: Scikit1D, predictions: Scikit1D) {
  *const labels = [1, 2, 3, 1]
   const predictions = [1, 2, 4, 4]
   let result = metrics.accuracyScore(labels, predictions)
-  console.log(result) // 0.5  
+  console.log(result) // 0.5
  *```
  * @param labels 1D Array-like that are the true values
  * @param predictions 1D Array-like that are your model predictions
