+++ shared/lib/pipeline/pipeline.test.ts
@@ -2,9 +2,9 @@ import { assert } from 'chai'
 import { Pipeline, makePipeline } from './pipeline'
 import { tensor1d } from '@tensorflow/tfjs-core'
 import { tensorEqual } from '../utils'
-import { LinearRegression } from '../estimators/linear.regression'
-import { SimpleImputer } from '../impute/simple.imputer'
-import { MinMaxScaler } from '../preprocessing/min.max.scaler'
+import { LinearRegression } from '../estimators/linearRegression'
+import { SimpleImputer } from '../impute/simpleImputer'
+import { MinMaxScaler } from '../preprocessing/minMaxScaler'
 import { describe, it } from 'mocha'
 
 describe('Pipeline', function () {
