+++ shared/lib/preprocessing/maxAbsScaler.ts
