+++ shared/lib/preprocessing/minMaxScaler.test.ts
@@ -1,5 +1,5 @@
 import { assert } from 'chai'
-import { MinMaxScaler } from './min.max.scaler'
+import { MinMaxScaler } from './minMaxScaler'
 import { dfd } from '../../globals'
 import { tensor2d } from '@tensorflow/tfjs-core'
 import { convertTensorToInputType } from '../utils'
