+++ shared/lib/preprocessing/minMaxScaler.ts
